+++ ml_tools/__init__.py
@@ -1,1 +1,2 @@
 from .custom_logger import custom_logger
+from ._schema import FeatureSchema